Add custom swagger UI endpoint
Allow to use the session cookie created when login via the web UI.
No need to implement login via Authorize in Swagger UI.

We load some custom javascript to inject a dummy bearer token that is required
by most API endpoints.
The session cookie is httponly and can't be retrieved from javascript.
The token is only used to force the app to check the session to see if the user is logged in or not.